+++ Dingo/Classes/CustomClasses/ZSTextField.m
@@ -183,7 +183,7 @@ UITableViewController *tableViewController;
 
 -(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    int count = [[self applyFilterWithSearchQuery:self.text] count];
+    NSInteger count = [[self applyFilterWithSearchQuery:self.text] count];
     if (count == 0) {
         [UIView animateWithDuration:0.3
                          animations:^{
